+++ tests/test_result.py
@@ -1,6 +1,11 @@
 from __future__ import annotations
+import sys
 
 from typing import Callable
+if sys.version_info >= (3, 10):
+    from typing import Literal
+else:
+    from typing_extensions import Literal
 
 import pytest
 
@@ -79,7 +84,7 @@ def test_err_method() -> None:
 
 
 def test_no_arg_ok() -> None:
-    top_level: Result[None, None] = Ok()
+    top_level: Result[Literal[True], None] = Ok(True)
     assert top_level.is_ok() is True
     assert top_level.ok() is True