+++ src/result/__init__.py
@@ -9,4 +9,4 @@ __all__ = [
     "as_async_result",
     "as_result",
 ]
-__version__ = "0.11.0.dev0"
+__version__ = "0.11.0rc1"