+++ CHANGELOG.md
@@ -12,6 +12,12 @@ Possible log types:
 - `[security]` to invite users to upgrade in case of vulnerabilities.
 
 
+## Unreleased
+
+- `[changed]` `Ok` now requires an explicit value during instantiation. Please
+  check out [MIGRATING.md], it will guide you through the necessary change in
+  your codebase.
+
 ## [0.10.0] - 2023-04-29
 
 - `[fixed]` Make python version check PEP 484 compliant (#118)
@@ -55,7 +61,6 @@ safe. Unfortunately this means some breaking changes. Please check out
 [MIGRATING.md], it will guide you through the necessary changes in your
 codebase.
 
-[MIGRATING.md]: https://github.com/rustedpy/result/blob/master/MIGRATING.md
 
 - [changed] Split result type into `Ok` and `Err` classes (#17, #27)
 - [deprecated] Python 3.4 support is deprecated and will be removed in the next
@@ -100,6 +105,7 @@ codebase.
 
  - Initial version
 
+[MIGRATING.md]: https://github.com/rustedpy/result/blob/master/MIGRATING.md
 [Unreleased]: https://github.com/rustedpy/result/compare/v0.10.0...HEAD
 [0.10.0]: https://github.com/rustedpy/result/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/rustedpy/result/compare/v0.8.0...v0.9.0