+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Possible log types:
 - `[security]` to invite users to upgrade in case of vulnerabilities.
 
 
-## Unreleased
+## [0.11.0] - 2023-06-11
 
 - `[changed]` `Ok` now requires an explicit value during instantiation. Please
   check out [MIGRATING.md], it will guide you through the necessary change in
@@ -108,7 +108,8 @@ codebase.
  - Initial version
 
 [MIGRATING.md]: https://github.com/rustedpy/result/blob/master/MIGRATING.md
-[Unreleased]: https://github.com/rustedpy/result/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/rustedpy/result/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/rustedpy/result/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/rustedpy/result/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/rustedpy/result/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/rustedpy/result/compare/v0.7.0...v0.8.0